+++ NEWS.md
@@ -1,6 +1,6 @@
 # rsample (development version)
 
-* `group_vfold_cv()` now supports stratification, so long as `balance = "groups"` (the default). Strata must be constant within each group (#317).
+* `group_vfold_cv()` now supports stratification. Strata must be constant within each group (#317, #360, #363, #364).
 
 * `group_bootstraps()` now warns if resampling returns any empty assessment sets (previously had been an error) (#356) (#357).